Changing filenames for carouselCard attributes
