+++ app/assets/main/react_components/common/MonthlyAction.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import CarouselActionItem from "../dashboard/climate_actions/carousel/CarouselActionItem.jsx";
+import CarouselCard from "../dashboard/climate_actions/carousel/CarouselCard.jsx";
 import { useClimateActionsText } from "../contexts/TextContext.js";
 import TextBanner from "./TextBanner.jsx";
 
@@ -8,15 +8,14 @@ const MonthlyAction = ({ action, user, updateLocalAccepted, categories }) => {
   const text = climateActionsText.monthly_action;
   return (
     <div className="w-64 ml-10 mt-11 relative">
-      <CarouselActionItem
+      <CarouselCard
         action={action}
         key={action.id}
         user={user}
         updateLocalAccepted={updateLocalAccepted}
         categories={categories}
         monthlyActionBanner={true}
-      ></CarouselActionItem>
-
+      ></CarouselCard>
       <div className="h-52 w-52 -top-4 left-6 bg-cover absolute">
         <TextBanner text={text} />
       </div>
