+++ app/assets/main/react_components/dashboard/climate_actions/carousel/carousel_card_attributes/CategoryColor.jsx
@@ -1,5 +1,6 @@
 import React from "react";
-const ShowCategoryColor = ({ categories, action }) => {
+
+const CategoryColor = ({ categories, action }) => {
   const categoryName = () => {
     for (let i = 0; i <= Object.keys(categories).length; i++) {
       if (categories[i].id === action.climate_action_category_id) {
@@ -18,4 +19,4 @@ const ShowCategoryColor = ({ categories, action }) => {
     ></div>
   );
 };
-export default ShowCategoryColor;
+export default CategoryColor;
